pyproject.toml
[tool.poetry]
name = "watchtower"
version = "0.1.0"
description = "Sentinels for some concerns"
authors = ["Dooon"]
license = "GNU GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
httpx = "^0.16.1"
selectolax = "^0.2.10"
python-dotenv = "^0.15.0"
furl = "^2.1.0"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
mypy = "^0.800"
isort = "^5.7.0"
flake8 = "^3.8.4"
pytest = "^6.2.2"
ipython = "^7.20.0"
pre-commit = "^2.10.1"
autoflake = "^1.4"
pytest-cov = "^2.11.1"
pytest-httpx = "^0.10.1"

[tool.black]
line-length = 119
target-version = ['py38']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | \.python-version
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
